requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
tensorflow==2.19.0
pillow
